backend/requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
python-dotenv
bcrypt